+++ CHANGELOG.md
@@ -6,14 +6,12 @@ FEATURES:
 
 - Add Kubelet Image GC support for SKS nodepools
 - Block storage volume resource & data source #341
+- Block storage volume snapshot resource & data source #344
 
 BUG FIXES:
 
 - Fix dbaas bugs causing accepteance tests to fail #346
 - docs: fix example in index.md #345 
-
-BUG FIXES:
-
 - Set labels on unmanaged eip creation #347 
 
 ## 0.57.0 (April 3, 2024)